requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
nepali-datetime
apscheduler
pywebview
pillow
